requirements.txt
streamlit
sentence-transformers
faiss-cpu
numpy
groq